requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
chromadb
sentence-transformers
groq
python-dotenv